refactor: separate RPC API services into rpc-live and rpc-stats
