+++ src/pages/network/assets/tokens/TokensPage.tsx
@@ -5,7 +5,7 @@ import { withHelmet } from '@app/components/hocs'
 import { Breadcrumbs } from '@app/components/ui'
 import { PageLayout } from '@app/components/ui/layouts'
 import { useTailwindBreakpoint } from '@app/hooks'
-import { useGetAssetsIssuancesQuery } from '@app/store/apis/archiver-v1'
+import { useGetAssetsIssuancesQuery } from '@app/store/apis/rpc-live'
 import type { TokenCategory } from '@app/store/apis/qubic-static'
 import { useGetTokenCategoriesQuery } from '@app/store/apis/qubic-static'
 import { ASSETS_ISSUER_ADDRESS } from '@app/utils/qubic-ts'
