+++ src/store/index.ts
@@ -4,6 +4,8 @@ import { archiverV1Api } from './apis/archiver-v1'
 import { archiverV2Api } from './apis/archiver-v2'
 import { qliApi } from './apis/qli'
 import { qubicStaticApi } from './apis/qubic-static'
+import { rpcLiveApi } from './apis/rpc-live'
+import { rpcStatsApi } from './apis/rpc-stats'
 import { testnetQubicRpcV1 } from './apis/testnet-qubic-rpc-v1'
 import localeReducer from './localeSlice'
 import searchReducer from './searchSlice'
@@ -15,6 +17,8 @@ export const store = configureStore({
     [qliApi.reducerPath]: qliApi.reducer,
     [archiverV1Api.reducerPath]: archiverV1Api.reducer,
     [archiverV2Api.reducerPath]: archiverV2Api.reducer,
+    [rpcLiveApi.reducerPath]: rpcLiveApi.reducer,
+    [rpcStatsApi.reducerPath]: rpcStatsApi.reducer,
     [testnetQubicRpcV1.reducerPath]: testnetQubicRpcV1.reducer,
     [qubicStaticApi.reducerPath]: qubicStaticApi.reducer
   },
@@ -24,6 +28,8 @@ export const store = configureStore({
       .concat(qliApi.middleware)
       .concat(archiverV1Api.middleware)
       .concat(archiverV2Api.middleware)
+      .concat(rpcLiveApi.middleware)
+      .concat(rpcStatsApi.middleware)
       .concat(testnetQubicRpcV1.middleware)
       .concat(qubicStaticApi.middleware)
 })
