+++ src/store/searchSlice.ts
@@ -3,9 +3,11 @@ import { createAsyncThunk, createSlice } from '@reduxjs/toolkit'
 
 import type { AppDispatch, RootState } from '@app/store'
 import type { TransactionWithType } from '@app/types'
-import type { GetAddressBalancesResponse, GetTickDataResponse } from './apis/archiver-v1'
+import type { GetTickDataResponse } from './apis/archiver-v1'
 import { archiverV1Api } from './apis/archiver-v1'
 import { archiverV2Api } from './apis/archiver-v2'
+import type { GetAddressBalancesResponse } from './apis/rpc-live'
+import { rpcLiveApi } from './apis/rpc-live'
 
 type HandlerResponse = GetAddressBalancesResponse | GetTickDataResponse | TransactionWithType
 
@@ -43,7 +45,7 @@ const makeSearchHandlers = (
   },
   [SearchType.ADDRESS]: async (query) => {
     const balance = await dispatch(
-      archiverV1Api.endpoints.getAddressBalances.initiate({ address: query })
+      rpcLiveApi.endpoints.getAddressBalances.initiate({ address: query })
     ).unwrap()
     return { balance }
   },
